% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_package_repos.R
\name{add_package_info}
\alias{add_package_info}
\title{Add R Package Information}
\usage{
add_package_info(df)
}
\arguments{
\item{df}{A data.frame from \code{\link{r_package_repos}}}
}
\value{
A data.frame with additional package information
}
\description{
Add R Package Information
}
