% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dcf.R
\name{download_dcf}
\alias{download_dcf}
\title{Download DESCRIPTION files}
\usage{
download_dcf(urls)
}
\arguments{
\item{urls}{URLs for description files}

\item{...}{additional to arguments to \code{GET}}
}
\value{
A list of the files and the description objects
}
\description{
Download DESCRIPTION files
}
