% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_table.R
\name{package_table}
\alias{package_table}
\title{Package Table for Printing}
\usage{
package_table(df)
}
\arguments{
\item{df}{A \code{data.frame} from \code{r_package_repos} that has
been subset to what you want}
}
\value{
A \code{data.frame} that works well with \code{kable}
}
\description{
Package Table for Printing
}
