% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_package_repos.R
\name{r_package_repos}
\alias{r_package_repos}
\title{Package Repository Information}
\usage{
r_package_repos(username = "muschellij2", ...)
}
\arguments{
\item{username}{GitHub username}

\item{...}{additional to arguments to \code{get_all_repos}}
}
\value{
A nested \code{data.frame} of repository information
}
\description{
Grabs all repository information from GitHub
}
