% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_has_file.R
\name{repo_has_file}
\alias{repo_has_file}
\alias{repo_has_description}
\title{Determine if GitHub Repository Has a file}
\usage{
repo_has_file(repo, file = "DESCRIPTION")

repo_has_description(repo)
}
\arguments{
\item{repo}{Github repository in the order or user/repo}

\item{file}{file to check if exists}
}
\value{
A character
}
\description{
Determine if GitHub Repository Has a file
}
\examples{
repo_has_file("muschellij2/fslr", "DESCRIPTION")
}
