% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis_svg_text.R
\name{travis_svg_status}
\alias{travis_svg_status}
\title{Status of Build from Travis SVG}
\usage{
travis_svg_status(url)
}
\arguments{
\item{url}{URL(s) of SVG badges from Travis}
}
\value{
A character vector build statuses.
}
\description{
Status of Build from Travis SVG
}
\examples{
url = "https://travis-ci.org/muschellij2/aal.svg?branch=master"
travis_svg_status(url)
# use .com vs. .org
com_url = "https://travis-ci.com/muschellij2/aal.svg?branch=master"
travis_svg_status(com_url)

}
