% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_contents.R
\name{repo_contents}
\alias{repo_contents}
\alias{repo_files}
\title{Get Repository Contents}
\usage{
repo_contents(repo, ...)

repo_files(repo, ...)
}
\arguments{
\item{repo}{Github repository in the order or user/repo}

\item{...}{arguments passed to \code{\link{gh}}}
}
\value{
A list or a character vector of files
}
\description{
Retrieves Github Repository Contents/filenames
}
\examples{
repo_contents("muschellij2/fslr")
}
